fix: 머지 충돌 해결

+++ app/src/main/java/com/example/hampouch/ui/mypage/MyPageScreen.kt
@@ -75,7 +75,7 @@ fun MyPageScreen(
     onNavigateToChallengeExpenseAnalysis: (totalDays: Int, periodStart: LocalDate, periodEnd: LocalDate) -> Unit = { _, _, _ -> },
     onNavigateToAmountAdjustment: () -> Unit = {},
     onNavigateToTakeABreak: () -> Unit = {},
-    onStartNewChallengeClick: () -> Unit = {}
+    onStartNewChallengeClick: (Int) -> Unit = {}
 ) {
     val context = LocalContext.current
     var route by rememberSaveable { mutableStateOf(MyPageRoute.MAIN) }
